New Home Screen and QSettings
No longer us draeer in the far corner to go between modes. Have a picture background with a orange start button and smaller buttons on the side. The volume is working and saving between runs. YAY! QSettings is working for volume - still need to do it for the numCoins though (just written the variable in the class didnt do anything with it)