Missions yay!
The timer now accounts for dates (not tested). Missions with the mission button should work, the model does save.  Spawns 3 random missions which save beteeen runs and should update if the a new day (even if offline). The list model info is saved and put between runs. The exclamantion mark thing is broken. Trying to bind to checkifButtonSHouldBeVisble() method.