Halftime Mode Begun
Okay so its been a while so im not sure what the last changes were. But now we have the startings of a half time mode, which is nice. fyi felgo isnt working yet. halfmode is not integrated in any way what so every yet. tbh im kindogg afraid of changing the compeititive mode and messig everything up. so that still needs to happen