More fixes
around 8 billion bugs fixed. added the last mission which changes each time and i think it works. font sizes and types were changed. this works right now so thats good. if needed shouldnt be to bad to revert to this. the last three missionmodel things have not been finished yet in implementation though. need to fix ads process for continue